requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Admin==1.6.2
Flask-Security-Too==5.1.5
Flask-Babel==3.1.0
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
WTForms==3.1.1
MarkupSafe==2.1.3